Merged in gh/stop-wait-for-instances (pull request #15)
flexctl stop -d should wait for workflow instances to finish before stopping the workflow.

* Implemented STOP deployment logic. When a STOP is issued, healthd
will monitor any running instances (those that are not in state
COMPLETED or ERROR) until they transition to those states. Then,
the deployment is destroyed as before.

This change does not use the wait_for_prefix() mechanism, because I
couldn't get that to work, but will get us past demo.

This change also adds a role for the healthd pod, and installs
both k8s and istio on the healthd pod so it can be used for
workflow manipulation.