fix coroutine object is not iterable "error"
It seems that a pythong coroutine object is being passed as data and
attempting to pass this on in an HTTP call causes problems. It doesn't
seem that this data element is actually useful, so removing it fixes
the issue, but we need more investigation to understand why a
coroutine object is there.