Merged in ui-bridge-fix-race-cond (pull request #187)
Fix ui-bridge race condition

* Fix ui-bridge race condition

We call flowd here to run the workflow, but if the first event in the
workflow is a user task, we will get a call back into app.init_route()
before we return here. The implication is that we don't know what the
iid is yet, as flowd has not returned, and our caller doesn't have
the iid (as we haven't returned it to them yet)

graphql_handlers.mutation_create_instance()
-> workflow.create_instance()
   -> flowd.RunWorkflow()
      -> catchgateway.create_instance()
         -> executor()
            -> catchgateway.run_instance() <= via executor
               -> app.init_route() # BOOM! iid is unknown
      -> return iid
   -> return iid
-> return iid <= caller learns IID at this point

Guard the iid/graphql_uri mapping with a semaphore to prevent this.